Cleans up SolutionView.jsx
