+++ src/tabs/AOCTab/days/day1/Day1Script.js
@@ -1,14 +1,16 @@
-const Day1Solution = (props) => {
-  const input = props.input
-  console.log("Day1 input: " + input); 
- 
-  const answer = "Day 1 answer";
+// Create component that returns the day1script.txt file
+//   OR create component that returns the string of the script
+
+// Think about what component should return
+
+const Day1Script = () => {
+  const answer = "./days/day1/day1script.txt";
   
   return(
-    <div> 
+    <>  
       {answer}
-    </div>
+    </>
   );
 }
 
-export default Day1Solution;
+export default Day1Script;
