+++ src/tabs/AOCTab/days/day1/Day1Solution.js
@@ -1,0 +1,14 @@
+const Day1Solution = (props) => {
+  const input = props.input
+  console.log("Day1 input: " + input); 
+ 
+  const answer = "Day 1 answer";
+  
+  return(
+    <div> 
+      {answer}
+    </div>
+  );
+}
+
+export default Day1Solution;
