+++ src/tabs/AOCTab/days/day2/Day2Solution.js
@@ -1,0 +1,13 @@
+const Day2Solution = (props) => {
+  const input = props.input
+  console.log("Day 2 input: " + input); 
+  
+  const answer = "Day 2 answer";
+  return(
+    <div>
+      {answer}
+    </div>
+  );
+}
+
+export default Day2Solution;
