--- src/tabs/AOCTab/days/day3/Day3Script.js
@@ -1,14 +1,0 @@
-const Day3Solution = (props) => {
-  const input = props.input
-  console.log("Day 3 input: " + input); 
-  
-  const answer = "Day 3 answer";
-
-  return(
-    <div>
-      {answer}
-    </div>
-  );
-}
-
-export default Day3Solution
