+++ src/tabs/AOCTab/days/day3/Day3Solution.js
@@ -1,0 +1,14 @@
+const Day3Solution = (props) => {
+  const input = props.input
+  console.log("Day 3 input: " + input); 
+  
+  const answer = "Day 3 answer";
+
+  return(
+    <div>
+      {answer}
+    </div>
+  );
+}
+
+export default Day3Solution
