--- src/tabs/AOCTab/days/day4/Day4Script.js
@@ -1,14 +1,0 @@
-const Day4Solution = (props) => {
-  const input = props.input
-  console.log("Day 4 input: " + input); 
-  
-  const answer = "Day 4 answer"; 
-  
-  return(
-    <div>
-      {answer}
-    </div>
-  );
-}
-
-export default Day4Solution;
