+++ src/tabs/AOCTab/days/day4/Day4Solution.js
@@ -1,0 +1,14 @@
+const Day4Solution = (props) => {
+  const input = props.input
+  console.log("Day 4 input: " + input); 
+  
+  const answer = "Day 4 answer"; 
+  
+  return(
+    <div>
+      {answer}
+    </div>
+  );
+}
+
+export default Day4Solution;
